requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.6.4
ta==0.10.2
flask==3.0.0
pyinstaller==6.3.0
plotly==5.18.0
scipy==1.11.4
scikit-learn==1.3.2
pdfkit==1.0.0 